services/voice_workflow/requirements.txt
fastapi
uvicorn
pydantic
openai-whisper
python-multipart 